+++ src/freeclimb/Api/DefaultApi.cs
@@ -1039,9 +1039,16 @@ namespace freeclimb.Api
         /// <exception cref="freeclimb.Client.ApiException">Thrown when fails to make API call</exception>
         /// <param name="callId">String that uniquely identifies this call resource.</param>
         /// <param name="dateCreated">Only show recordings created on the specified date, in the form *YYYY-MM-DD*. (optional)</param>
+        /// <param name="startTime">Only show Recordings created at or after this time, given as YYYY-MM-DD hh:mm:ss. (optional)</param>
+        /// <param name="endTime">Only show Recordings created at or before this time, given as YYYY-MM-DD hh:mm:ss. (optional)</param>
 
         /// <returns>RecordingList</returns>
-        RecordingList ListCallRecordings(string callId, string? dateCreated = default(string?));
+        RecordingList ListCallRecordings(
+            string callId,
+            string? dateCreated = default(string?),
+            string? startTime = default(string?),
+            string? endTime = default(string?)
+        );
 
         /// <summary>
         /// List Call Recordings
@@ -1052,17 +1059,22 @@ namespace freeclimb.Api
         /// <exception cref="freeclimb.Client.ApiException">Thrown when fails to make API call</exception>
         /// <param name="callId">String that uniquely identifies this call resource.</param>
         /// <param name="dateCreated">Only show recordings created on the specified date, in the form *YYYY-MM-DD*. (optional)</param>
+        /// <param name="startTime">Only show Recordings created at or after this time, given as YYYY-MM-DD hh:mm:ss. (optional)</param>
+        /// <param name="endTime">Only show Recordings created at or before this time, given as YYYY-MM-DD hh:mm:ss. (optional)</param>
 
         /// <returns>ApiResponse of RecordingList</returns>
         ApiResponse<RecordingList> ListCallRecordingsWithHttpInfo(
             string callId,
-            string? dateCreated = default(string?)
+            string? dateCreated = default(string?),
+            string? startTime = default(string?),
+            string? endTime = default(string?)
         );
 
         /// <summary>
         /// List Calls
         /// </summary>
         /// <exception cref="freeclimb.Client.ApiException">Thrown when fails to make API call</exception>
+        /// <param name="usedAudioStream">If usedAudioStream is set to true then all calls that have a audioStreamDuration &gt; 0 will be returned  (optional, default to false)</param>
         /// <param name="active">If active is set to true then all calls of the nature queued, ringing, inProgress are returned in the query. (optional, default to false)</param>
         /// <param name="to">Only show Calls to this phone number. (optional)</param>
         /// <param name="from">Only show Calls from this phone number. (optional)</param>
@@ -1073,9 +1085,11 @@ namespace freeclimb.Api
         /// <param name="applicationId">Only show calls belonging to the given applicationId. This parameter can be repeated to return calls from multiple Applications. (optional)</param>
         /// <param name="riskScoreMin">The minimum riskScore that should be included in the list. (optional)</param>
         /// <param name="riskScoreMax">The maximum riskScore that should be included in the list. (optional)</param>
+        /// <param name="webRTC">Only show Calls that were originated via WebRTC. (optional, default to false)</param>
 
         /// <returns>CallList</returns>
         CallList ListCalls(
+            bool? usedAudioStream = default(bool?),
             bool? active = default(bool?),
             string? to = default(string?),
             string? from = default(string?),
@@ -1085,7 +1099,8 @@ namespace freeclimb.Api
             string? parentCallId = default(string?),
             List<string>? applicationId = default(List<string>?),
             int? riskScoreMin = default(int?),
-            int? riskScoreMax = default(int?)
+            int? riskScoreMax = default(int?),
+            bool? webRTC = default(bool?)
         );
 
         /// <summary>
@@ -1095,6 +1110,7 @@ namespace freeclimb.Api
         ///
         /// </remarks>
         /// <exception cref="freeclimb.Client.ApiException">Thrown when fails to make API call</exception>
+        /// <param name="usedAudioStream">If usedAudioStream is set to true then all calls that have a audioStreamDuration &gt; 0 will be returned  (optional, default to false)</param>
         /// <param name="active">If active is set to true then all calls of the nature queued, ringing, inProgress are returned in the query. (optional, default to false)</param>
         /// <param name="to">Only show Calls to this phone number. (optional)</param>
         /// <param name="from">Only show Calls from this phone number. (optional)</param>
@@ -1105,9 +1121,11 @@ namespace freeclimb.Api
         /// <param name="applicationId">Only show calls belonging to the given applicationId. This parameter can be repeated to return calls from multiple Applications. (optional)</param>
         /// <param name="riskScoreMin">The minimum riskScore that should be included in the list. (optional)</param>
         /// <param name="riskScoreMax">The maximum riskScore that should be included in the list. (optional)</param>
+        /// <param name="webRTC">Only show Calls that were originated via WebRTC. (optional, default to false)</param>
 
         /// <returns>ApiResponse of CallList</returns>
         ApiResponse<CallList> ListCallsWithHttpInfo(
+            bool? usedAudioStream = default(bool?),
             bool? active = default(bool?),
             string? to = default(string?),
             string? from = default(string?),
@@ -1117,7 +1135,8 @@ namespace freeclimb.Api
             string? parentCallId = default(string?),
             List<string>? applicationId = default(List<string>?),
             int? riskScoreMin = default(int?),
-            int? riskScoreMax = default(int?)
+            int? riskScoreMax = default(int?),
+            bool? webRTC = default(bool?)
         );
 
         /// <summary>
@@ -1127,12 +1146,16 @@ namespace freeclimb.Api
         /// <param name="conferenceId">Show only Recordings made during the conference with this ID.</param>
         /// <param name="callId">Show only Recordings made during the Call with this ID. (optional)</param>
         /// <param name="dateCreated">Only show Recordings created on this date, formatted as *YYYY-MM-DD*. (optional)</param>
+        /// <param name="startTime">Only show Recordings created at or after this time, given as YYYY-MM-DD hh:mm:ss. (optional)</param>
+        /// <param name="endTime">Only show Recordings created at or before this time, given as YYYY-MM-DD hh:mm:ss. (optional)</param>
 
         /// <returns>RecordingList</returns>
         RecordingList ListConferenceRecordings(
             string conferenceId,
             string? callId = default(string?),
-            string? dateCreated = default(string?)
+            string? dateCreated = default(string?),
+            string? startTime = default(string?),
+            string? endTime = default(string?)
         );
 
         /// <summary>
@@ -1145,12 +1168,16 @@ namespace freeclimb.Api
         /// <param name="conferenceId">Show only Recordings made during the conference with this ID.</param>
         /// <param name="callId">Show only Recordings made during the Call with this ID. (optional)</param>
         /// <param name="dateCreated">Only show Recordings created on this date, formatted as *YYYY-MM-DD*. (optional)</param>
+        /// <param name="startTime">Only show Recordings created at or after this time, given as YYYY-MM-DD hh:mm:ss. (optional)</param>
+        /// <param name="endTime">Only show Recordings created at or before this time, given as YYYY-MM-DD hh:mm:ss. (optional)</param>
 
         /// <returns>ApiResponse of RecordingList</returns>
         ApiResponse<RecordingList> ListConferenceRecordingsWithHttpInfo(
             string conferenceId,
             string? callId = default(string?),
-            string? dateCreated = default(string?)
+            string? dateCreated = default(string?),
+            string? startTime = default(string?),
+            string? endTime = default(string?)
         );
 
         /// <summary>
@@ -1361,12 +1388,16 @@ namespace freeclimb.Api
         /// <param name="callId">Show only Recordings made during the Call with this ID. (optional)</param>
         /// <param name="conferenceId">Show only Recordings made during the conference with this ID. (optional)</param>
         /// <param name="dateCreated">Only show Recordings created on this date, formatted as *YYYY-MM-DD*. (optional)</param>
+        /// <param name="startTime">Only show Recordings created at or after this time, given as YYYY-MM-DD hh:mm:ss. (optional)</param>
+        /// <param name="endTime">Only show Recordings created at or before this time, given as YYYY-MM-DD hh:mm:ss. (optional)</param>
 
         /// <returns>RecordingList</returns>
         RecordingList ListRecordings(
             string? callId = default(string?),
             string? conferenceId = default(string?),
-            string? dateCreated = default(string?)
+            string? dateCreated = default(string?),
+            string? startTime = default(string?),
+            string? endTime = default(string?)
         );
 
         /// <summary>
@@ -1379,12 +1410,16 @@ namespace freeclimb.Api
         /// <param name="callId">Show only Recordings made during the Call with this ID. (optional)</param>
         /// <param name="conferenceId">Show only Recordings made during the conference with this ID. (optional)</param>
         /// <param name="dateCreated">Only show Recordings created on this date, formatted as *YYYY-MM-DD*. (optional)</param>
+        /// <param name="startTime">Only show Recordings created at or after this time, given as YYYY-MM-DD hh:mm:ss. (optional)</param>
+        /// <param name="endTime">Only show Recordings created at or before this time, given as YYYY-MM-DD hh:mm:ss. (optional)</param>
 
         /// <returns>ApiResponse of RecordingList</returns>
         ApiResponse<RecordingList> ListRecordingsWithHttpInfo(
             string? callId = default(string?),
             string? conferenceId = default(string?),
-            string? dateCreated = default(string?)
+            string? dateCreated = default(string?),
+            string? startTime = default(string?),
+            string? endTime = default(string?)
         );
 
         /// <summary>
@@ -6935,15 +6970,19 @@ namespace freeclimb.Api
         /// <exception cref="freeclimb.Client.ApiException">Thrown when fails to make API call</exception>
         /// <param name="callId">String that uniquely identifies this call resource.</param>
         /// <param name="dateCreated">Only show recordings created on the specified date, in the form *YYYY-MM-DD*. (optional)</param>
+        /// <param name="startTime">Only show Recordings created at or after this time, given as YYYY-MM-DD hh:mm:ss. (optional)</param>
+        /// <param name="endTime">Only show Recordings created at or before this time, given as YYYY-MM-DD hh:mm:ss. (optional)</param>
 
         /// <returns>RecordingList</returns>
         public RecordingList ListCallRecordings(
             string callId,
-            string? dateCreated = default(string?)
+            string? dateCreated = default(string?),
+            string? startTime = default(string?),
+            string? endTime = default(string?)
         )
         {
             freeclimb.Client.ApiResponse<RecordingList> localVarResponse =
-                ListCallRecordingsWithHttpInfo(callId, dateCreated);
+                ListCallRecordingsWithHttpInfo(callId, dateCreated, startTime, endTime);
             return localVarResponse.Data;
         }
 
@@ -6953,11 +6992,15 @@ namespace freeclimb.Api
         /// <exception cref="freeclimb.Client.ApiException">Thrown when fails to make API call</exception>
         /// <param name="callId">String that uniquely identifies this call resource.</param>
         /// <param name="dateCreated">Only show recordings created on the specified date, in the form *YYYY-MM-DD*. (optional)</param>
+        /// <param name="startTime">Only show Recordings created at or after this time, given as YYYY-MM-DD hh:mm:ss. (optional)</param>
+        /// <param name="endTime">Only show Recordings created at or before this time, given as YYYY-MM-DD hh:mm:ss. (optional)</param>
 
         /// <returns>ApiResponse of RecordingList</returns>
         public freeclimb.Client.ApiResponse<RecordingList> ListCallRecordingsWithHttpInfo(
             string callId,
-            string? dateCreated = default(string?)
+            string? dateCreated = default(string?),
+            string? startTime = default(string?),
+            string? endTime = default(string?)
         )
         {
             // verify the required parameter 'accountId' is set
@@ -7014,6 +7057,18 @@ namespace freeclimb.Api
                     freeclimb.Client.ClientUtils.ParameterToMultiMap("", "dateCreated", dateCreated)
                 );
             }
+            if (startTime != null)
+            {
+                localVarRequestOptions.QueryParameters.Add(
+                    freeclimb.Client.ClientUtils.ParameterToMultiMap("", "startTime", startTime)
+                );
+            }
+            if (endTime != null)
+            {
+                localVarRequestOptions.QueryParameters.Add(
+                    freeclimb.Client.ClientUtils.ParameterToMultiMap("", "endTime", endTime)
+                );
+            }
 
             localVarRequestOptions.Operation = "DefaultApi.ListCallRecordings";
 
@@ -7059,6 +7114,7 @@ namespace freeclimb.Api
         /// List Calls
         /// </summary>
         /// <exception cref="freeclimb.Client.ApiException">Thrown when fails to make API call</exception>
+        /// <param name="usedAudioStream">If usedAudioStream is set to true then all calls that have a audioStreamDuration &gt; 0 will be returned  (optional, default to false)</param>
         /// <param name="active">If active is set to true then all calls of the nature queued, ringing, inProgress are returned in the query. (optional, default to false)</param>
         /// <param name="to">Only show Calls to this phone number. (optional)</param>
         /// <param name="from">Only show Calls from this phone number. (optional)</param>
@@ -7069,9 +7125,11 @@ namespace freeclimb.Api
         /// <param name="applicationId">Only show calls belonging to the given applicationId. This parameter can be repeated to return calls from multiple Applications. (optional)</param>
         /// <param name="riskScoreMin">The minimum riskScore that should be included in the list. (optional)</param>
         /// <param name="riskScoreMax">The maximum riskScore that should be included in the list. (optional)</param>
+        /// <param name="webRTC">Only show Calls that were originated via WebRTC. (optional, default to false)</param>
 
         /// <returns>CallList</returns>
         public CallList ListCalls(
+            bool? usedAudioStream = default(bool?),
             bool? active = default(bool?),
             string? to = default(string?),
             string? from = default(string?),
@@ -7081,10 +7139,12 @@ namespace freeclimb.Api
             string? parentCallId = default(string?),
             List<string>? applicationId = default(List<string>?),
             int? riskScoreMin = default(int?),
-            int? riskScoreMax = default(int?)
+            int? riskScoreMax = default(int?),
+            bool? webRTC = default(bool?)
         )
         {
             freeclimb.Client.ApiResponse<CallList> localVarResponse = ListCallsWithHttpInfo(
+                usedAudioStream,
                 active,
                 to,
                 from,
@@ -7094,7 +7154,8 @@ namespace freeclimb.Api
                 parentCallId,
                 applicationId,
                 riskScoreMin,
-                riskScoreMax
+                riskScoreMax,
+                webRTC
             );
             return localVarResponse.Data;
         }
@@ -7103,6 +7164,7 @@ namespace freeclimb.Api
         /// List Calls
         /// </summary>
         /// <exception cref="freeclimb.Client.ApiException">Thrown when fails to make API call</exception>
+        /// <param name="usedAudioStream">If usedAudioStream is set to true then all calls that have a audioStreamDuration &gt; 0 will be returned  (optional, default to false)</param>
         /// <param name="active">If active is set to true then all calls of the nature queued, ringing, inProgress are returned in the query. (optional, default to false)</param>
         /// <param name="to">Only show Calls to this phone number. (optional)</param>
         /// <param name="from">Only show Calls from this phone number. (optional)</param>
@@ -7113,9 +7175,11 @@ namespace freeclimb.Api
         /// <param name="applicationId">Only show calls belonging to the given applicationId. This parameter can be repeated to return calls from multiple Applications. (optional)</param>
         /// <param name="riskScoreMin">The minimum riskScore that should be included in the list. (optional)</param>
         /// <param name="riskScoreMax">The maximum riskScore that should be included in the list. (optional)</param>
+        /// <param name="webRTC">Only show Calls that were originated via WebRTC. (optional, default to false)</param>
 
         /// <returns>ApiResponse of CallList</returns>
         public freeclimb.Client.ApiResponse<CallList> ListCallsWithHttpInfo(
+            bool? usedAudioStream = default(bool?),
             bool? active = default(bool?),
             string? to = default(string?),
             string? from = default(string?),
@@ -7125,7 +7189,8 @@ namespace freeclimb.Api
             string? parentCallId = default(string?),
             List<string>? applicationId = default(List<string>?),
             int? riskScoreMin = default(int?),
-            int? riskScoreMax = default(int?)
+            int? riskScoreMax = default(int?),
+            bool? webRTC = default(bool?)
         )
         {
             // verify the required parameter 'accountId' is set
@@ -7163,6 +7228,16 @@ namespace freeclimb.Api
                 "accountId",
                 freeclimb.Client.ClientUtils.ParameterToString(accountId)
             ); // path parameter
+            if (usedAudioStream != null)
+            {
+                localVarRequestOptions.QueryParameters.Add(
+                    freeclimb.Client.ClientUtils.ParameterToMultiMap(
+                        "",
+                        "usedAudioStream",
+                        usedAudioStream
+                    )
+                );
+            }
             if (active != null)
             {
                 localVarRequestOptions.QueryParameters.Add(
@@ -7239,6 +7314,12 @@ namespace freeclimb.Api
                     )
                 );
             }
+            if (webRTC != null)
+            {
+                localVarRequestOptions.QueryParameters.Add(
+                    freeclimb.Client.ClientUtils.ParameterToMultiMap("", "webRTC", webRTC)
+                );
+            }
 
             localVarRequestOptions.Operation = "DefaultApi.ListCalls";
 
@@ -7284,16 +7365,26 @@ namespace freeclimb.Api
         /// <param name="conferenceId">Show only Recordings made during the conference with this ID.</param>
         /// <param name="callId">Show only Recordings made during the Call with this ID. (optional)</param>
         /// <param name="dateCreated">Only show Recordings created on this date, formatted as *YYYY-MM-DD*. (optional)</param>
+        /// <param name="startTime">Only show Recordings created at or after this time, given as YYYY-MM-DD hh:mm:ss. (optional)</param>
+        /// <param name="endTime">Only show Recordings created at or before this time, given as YYYY-MM-DD hh:mm:ss. (optional)</param>
 
         /// <returns>RecordingList</returns>
         public RecordingList ListConferenceRecordings(
             string conferenceId,
             string? callId = default(string?),
-            string? dateCreated = default(string?)
+            string? dateCreated = default(string?),
+            string? startTime = default(string?),
+            string? endTime = default(string?)
         )
         {
             freeclimb.Client.ApiResponse<RecordingList> localVarResponse =
-                ListConferenceRecordingsWithHttpInfo(conferenceId, callId, dateCreated);
+                ListConferenceRecordingsWithHttpInfo(
+                    conferenceId,
+                    callId,
+                    dateCreated,
+                    startTime,
+                    endTime
+                );
             return localVarResponse.Data;
         }
 
@@ -7304,12 +7395,16 @@ namespace freeclimb.Api
         /// <param name="conferenceId">Show only Recordings made during the conference with this ID.</param>
         /// <param name="callId">Show only Recordings made during the Call with this ID. (optional)</param>
         /// <param name="dateCreated">Only show Recordings created on this date, formatted as *YYYY-MM-DD*. (optional)</param>
+        /// <param name="startTime">Only show Recordings created at or after this time, given as YYYY-MM-DD hh:mm:ss. (optional)</param>
+        /// <param name="endTime">Only show Recordings created at or before this time, given as YYYY-MM-DD hh:mm:ss. (optional)</param>
 
         /// <returns>ApiResponse of RecordingList</returns>
         public freeclimb.Client.ApiResponse<RecordingList> ListConferenceRecordingsWithHttpInfo(
             string conferenceId,
             string? callId = default(string?),
-            string? dateCreated = default(string?)
+            string? dateCreated = default(string?),
+            string? startTime = default(string?),
+            string? endTime = default(string?)
         )
         {
             // verify the required parameter 'accountId' is set
@@ -7370,6 +7465,18 @@ namespace freeclimb.Api
             {
                 localVarRequestOptions.QueryParameters.Add(
                     freeclimb.Client.ClientUtils.ParameterToMultiMap("", "dateCreated", dateCreated)
+                );
+            }
+            if (startTime != null)
+            {
+                localVarRequestOptions.QueryParameters.Add(
+                    freeclimb.Client.ClientUtils.ParameterToMultiMap("", "startTime", startTime)
+                );
+            }
+            if (endTime != null)
+            {
+                localVarRequestOptions.QueryParameters.Add(
+                    freeclimb.Client.ClientUtils.ParameterToMultiMap("", "endTime", endTime)
                 );
             }
 
@@ -8216,16 +8323,20 @@ namespace freeclimb.Api
         /// <param name="callId">Show only Recordings made during the Call with this ID. (optional)</param>
         /// <param name="conferenceId">Show only Recordings made during the conference with this ID. (optional)</param>
         /// <param name="dateCreated">Only show Recordings created on this date, formatted as *YYYY-MM-DD*. (optional)</param>
+        /// <param name="startTime">Only show Recordings created at or after this time, given as YYYY-MM-DD hh:mm:ss. (optional)</param>
+        /// <param name="endTime">Only show Recordings created at or before this time, given as YYYY-MM-DD hh:mm:ss. (optional)</param>
 
         /// <returns>RecordingList</returns>
         public RecordingList ListRecordings(
             string? callId = default(string?),
             string? conferenceId = default(string?),
-            string? dateCreated = default(string?)
+            string? dateCreated = default(string?),
+            string? startTime = default(string?),
+            string? endTime = default(string?)
         )
         {
             freeclimb.Client.ApiResponse<RecordingList> localVarResponse =
-                ListRecordingsWithHttpInfo(callId, conferenceId, dateCreated);
+                ListRecordingsWithHttpInfo(callId, conferenceId, dateCreated, startTime, endTime);
             return localVarResponse.Data;
         }
 
@@ -8236,12 +8347,16 @@ namespace freeclimb.Api
         /// <param name="callId">Show only Recordings made during the Call with this ID. (optional)</param>
         /// <param name="conferenceId">Show only Recordings made during the conference with this ID. (optional)</param>
         /// <param name="dateCreated">Only show Recordings created on this date, formatted as *YYYY-MM-DD*. (optional)</param>
+        /// <param name="startTime">Only show Recordings created at or after this time, given as YYYY-MM-DD hh:mm:ss. (optional)</param>
+        /// <param name="endTime">Only show Recordings created at or before this time, given as YYYY-MM-DD hh:mm:ss. (optional)</param>
 
         /// <returns>ApiResponse of RecordingList</returns>
         public freeclimb.Client.ApiResponse<RecordingList> ListRecordingsWithHttpInfo(
             string? callId = default(string?),
             string? conferenceId = default(string?),
-            string? dateCreated = default(string?)
+            string? dateCreated = default(string?),
+            string? startTime = default(string?),
+            string? endTime = default(string?)
         )
         {
             // verify the required parameter 'accountId' is set
@@ -8299,6 +8414,18 @@ namespace freeclimb.Api
             {
                 localVarRequestOptions.QueryParameters.Add(
                     freeclimb.Client.ClientUtils.ParameterToMultiMap("", "dateCreated", dateCreated)
+                );
+            }
+            if (startTime != null)
+            {
+                localVarRequestOptions.QueryParameters.Add(
+                    freeclimb.Client.ClientUtils.ParameterToMultiMap("", "startTime", startTime)
+                );
+            }
+            if (endTime != null)
+            {
+                localVarRequestOptions.QueryParameters.Add(
+                    freeclimb.Client.ClientUtils.ParameterToMultiMap("", "endTime", endTime)
                 );
             }